requirements.txt
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
Flask-WTF==1.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
selenium==4.24.0
simple-websocket==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
zipp==3.20.2
